:monocle_face: minor fixes
