+++ api/src/main/java/com/trainlab/controller/RoleController.java
@@ -3,53 +3,107 @@ package com.trainlab.controller;
 import com.trainlab.dto.RoleCreateDto;
 import com.trainlab.dto.RoleDto;
 import com.trainlab.model.Role;
-import com.trainlab.service.RoleService;
-import io.swagger.v3.oas.annotations.Parameter;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-@RestController
-@RequestMapping("api/v1/roles")
-@RequiredArgsConstructor
-public class RoleController implements RoleApi {
+public interface RoleController {
+    @Operation(
+            summary = "Spring Data Create role",
+            description = "Creates a new role",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "CREATED",
+                            description = "Created a new role",
+                            content = @Content(
+                                    mediaType = "application/json",
+                                    schema = @Schema(implementation = Role.class)
+                            )
+                    ),
+                    @ApiResponse(
+                            responseCode = "BAD_REQUEST",
+                            description = "Validation error"
+                    )
+            }
+    )
+    ResponseEntity<RoleDto> createRole(RoleCreateDto roleCreateDto);
 
-    private final RoleService roleService;
+    @Operation(
+            summary = "Spring Data receiving all roles",
+            description = "Receiving all roles without limitations",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "OK",
+                            description = "All user roles have successfully loaded",
+                            content = @Content(
+                                    mediaType = "application/json",
+                                    array = @ArraySchema(schema = @Schema(implementation = Role.class)))
 
-    @PostMapping("/")
-    public ResponseEntity<RoleDto> createRole(@Valid @RequestBody @Parameter(description = "Role DTO", required = true) RoleCreateDto roleCreateDto) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(roleService.create(roleCreateDto));
-    }
+                    ),
+                    @ApiResponse(responseCode = "INTERNAL_SERVER_ERROR", description = "Internal Server Error")
+            }
+    )
+    ResponseEntity<List<RoleDto>> findAll();
 
-    @GetMapping
-    public ResponseEntity<List<RoleDto>> findAll() {
-        return ResponseEntity.status(HttpStatus.OK).body(roleService.findAll());
-    }
+    @Operation(
+            summary = "Spring Data Role searching by its id",
+            description = "Receive role by id",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "OK",
+                            description = "Role has loaded Successfully",
+                            content = @Content(
+                                    mediaType = "application/json",
+                                    schema = @Schema(implementation = Role.class))
+                    ),
+                    @ApiResponse(
+                            responseCode = "NOT_FOUND",
+                            description = "Role not found"
+                    )
+            }
+    )
+    ResponseEntity<RoleDto> findRoleById(Integer id);
 
-    @GetMapping("/{id}")
-    public ResponseEntity<RoleDto> findRoleById(@PathVariable Integer id) {
-        return ResponseEntity.status(HttpStatus.OK).body(roleService.findById(id));
-    }
+    @Operation(
+            summary = "Spring Data Role receiving by role name",
+            description = "Receive role by role name",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "OK",
+                            description = "Role has loaded Successfully",
+                            content = @Content(
+                                    mediaType = "application/json",
+                                    schema = @Schema(implementation = Role.class))
+                    ),
+                    @ApiResponse(
+                            responseCode = "NOT_FOUND",
+                            description = "Role not found"
+                    )
+            }
+    )
+    ResponseEntity<RoleDto> findRoleByRoleName(String roleName);
 
-    @GetMapping("/")
-    public ResponseEntity<RoleDto> findRoleByRoleName(@RequestBody String roleName) {
-        return ResponseEntity.status(HttpStatus.OK).body(roleService.findByRoleName(roleName));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<RoleDto> updateRole(@PathVariable("id") Integer id,
-                                           @Valid @RequestBody RoleDto roleDto) {
-        return ResponseEntity.status(HttpStatus.OK).body(roleService.update(roleDto, id));
-    }
-
+    @Operation(
+            summary = "Spring Data Update role ",
+            description = "Updating based user role on given id and request body",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "OK",
+                            description = "User role has successfully updated",
+                            content = @Content(
+                                    mediaType = "application/json",
+                                    schema = @Schema(implementation = Role.class))
+                    ),
+                    @ApiResponse(
+                            responseCode = "BAD_REQUEST",
+                            description = "Validation error"
+                    )
+            }
+    )
+    ResponseEntity<RoleDto> updateRole(Integer id, RoleDto roleDto);
 }
