+++ common/src/main/java/com/trainlab/dto/RoleDto.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.sql.Timestamp;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -15,5 +17,11 @@ public class RoleDto {
     @Size(max = 30)
     @NotNull
     private String roleName;
+
+    @NotNull
+    private Timestamp created;
+
+    @NotNull
+    private Timestamp changed;
 
 }
